Successfully separated functions!
