+++ config.py
@@ -1,11 +1,14 @@
+### GLOBAL VARIABLE MODULE ###
+
 import gspread
 from oauth2client.service_account import ServiceAccountCredentials
-from tkinter import *
-from tkinter import ttk
 
+# GLOBAL Variables that change with the text boxes
 CELL_COL = "D"
 CELL_CONTENT = "1"
 GOOGLE_SHEETS_FILENAME = "Copy of 20-21 Points Sheet Template (Color Change)"
+
+
 # use creds to create a client to interact with the Google Drive API
 scope = ['https://spreadsheets.google.com/feeds',
          'https://www.googleapis.com/auth/drive']
